Added screenshot for stock transaction entry and stock_shares_new_stock.png
Added a new screenshot  showing the stock transaction entry interface.
This image is used for the Italian documentation to illustrate the process of adding stock transactions in MMEX.
Ensured consistency with the previously updated translation.

Added stock_shares_new_stock.png for Italian documentation. Uploaded an image for the Italian Stock Portfolio documentation.